+++ tests/SqidsBlocklistTest.php
@@ -21,74 +21,89 @@ class SqidsBlocklistTest extends TestCase
     {
         $sqids = new Sqids();
 
-        $this->assertSame([200044], $sqids->decode('sexy'));
-        $this->assertSame('d171vI', $sqids->encode([200044]));
+        $this->assertSame([4572721], $sqids->decode('aho1e'));
+        $this->assertSame('JExTR', $sqids->encode([4572721]));
     }
 
     public function testIfEmptyBlocklistParamPassedDontUseAnyBlocklist()
     {
         $sqids = new Sqids('', 0, []);
 
-        $this->assertSame([200044], $sqids->decode('sexy'));
-        $this->assertSame('sexy', $sqids->encode([200044]));
+        $this->assertSame([4572721], $sqids->decode('aho1e'));
+        $this->assertSame('aho1e', $sqids->encode([4572721]));
     }
 
     public function testIfNonEmptyBlocklistParamPassedUseOnlyThat()
     {
         $sqids = new Sqids('', 0, [
-            'AvTg' // originally encoded [100000]
+            'ArUO' // originally encoded [100000]
         ]);
 
         // Make sure we don't use the default blocklist
-        $this->assertSame([200044], $sqids->decode('sexy'));
-        $this->assertSame('sexy', $sqids->encode([200044]));
+        $this->assertSame([4572721], $sqids->decode('aho1e'));
+        $this->assertSame('aho1e', $sqids->encode([4572721]));
 
         // Make sure we are using the passed blocklist
-        $this->assertSame([100000], $sqids->decode('AvTg'));
-        $this->assertSame('7T1X8k', $sqids->encode([100000]));
-        $this->assertSame([100000], $sqids->decode('7T1X8k'));
+        $this->assertSame([100000], $sqids->decode('ArUO'));
+        $this->assertSame('QyG4', $sqids->encode([100000]));
+        $this->assertSame([100000], $sqids->decode('QyG4'));
     }
 
     public function testBlocklist()
     {
         $sqids = new Sqids('', 0, [
-            '8QRLaD', // normal result of 1st encoding, let's block that word on purpose
-            '7T1cd0dL', // result of 2nd encoding
-            'UeIe', // result of 3rd encoding is `RA8UeIe7`, let's block a substring
-            'imhw', // result of 4th encoding is `WM3Limhw`, let's block the postfix
-            'LfUQ' // result of 4th encoding is `LfUQh4HN`, let's block the prefix
+            'JSwXFaosAN', // normal result of 1st encoding, let's block that word on purpose
+            'OCjV9JK64o', // result of 2nd encoding
+            'rBHf', // result of 3rd encoding is `4rBHfOiqd3`, let's block a substring
+            '79SM', // result of 4th encoding is `dyhgw479SM`, let's block the postfix
+            '7tE6', // result of 4th encoding is `7tE6jdAHLe`, let's block the prefix
         ]);
 
-        $this->assertSame('TM0x1Mxz', $sqids->encode([1, 2, 3]));
-        $this->assertSame([1, 2, 3], $sqids->decode('TM0x1Mxz'));
+        $this->assertSame('1aYeB7bRUt', $sqids->encode([1_000_000, 2_000_000]));
+        $this->assertSame([1_000_000, 2_000_000], $sqids->decode('1aYeB7bRUt'));
     }
 
     public function testDecodingBlocklistWordsShouldStillWork()
     {
-        $sqids = new Sqids('', 0, ['8QRLaD', '7T1cd0dL', 'RA8UeIe7', 'WM3Limhw', 'LfUQh4HN']);
+        $sqids = new Sqids('', 0, ['86Rf07', 'se8ojk', 'ARsz1p', 'Q8AI49', '5sQRZO']);
 
-        $this->assertSame([1, 2, 3], $sqids->decode('8QRLaD'));
-        $this->assertSame([1, 2, 3], $sqids->decode('7T1cd0dL'));
-        $this->assertSame([1, 2, 3], $sqids->decode('RA8UeIe7'));
-        $this->assertSame([1, 2, 3], $sqids->decode('WM3Limhw'));
-        $this->assertSame([1, 2, 3], $sqids->decode('LfUQh4HN'));
+        $this->assertSame([1, 2, 3], $sqids->decode('86Rf07'));
+        $this->assertSame([1, 2, 3], $sqids->decode('se8ojk'));
+        $this->assertSame([1, 2, 3], $sqids->decode('ARsz1p'));
+        $this->assertSame([1, 2, 3], $sqids->decode('Q8AI49'));
+        $this->assertSame([1, 2, 3], $sqids->decode('5sQRZO'));
     }
 
     public function testMatchAgainstAShortBlocklistWord()
     {
-        $sqids = new Sqids('', 0, ['pPQ']);
+        $sqids = new Sqids('', 0, ['pnd']);
         $this->assertSame([1000], $sqids->decode($sqids->encode([1000])));
     }
 
     public function testBlocklistFilteringInConstructor()
     {
         // lowercase blocklist in only-uppercase alphabet
-        $sqids = new Sqids('ABCDEFGHIJKLMNOPQRSTUVWXYZ', 0, ['sqnmpn']);
+        $sqids = new Sqids('ABCDEFGHIJKLMNOPQRSTUVWXYZ', 0, ['sxnzkl']);
 
         $id = $sqids->encode([1, 2, 3]);
         $numbers = $sqids->decode($id);
 
-        $this->assertSame($id, 'ULPBZGBM'); // without blocklist, would've been "SQNMPN"
+        $this->assertSame($id, 'IBSHOZ'); // without blocklist, would've been "SXNZKL"
         $this->assertSame($numbers, [1, 2, 3]);
+    }
+
+    public function testMaxEncodingAttempts()
+    {
+        $alphabet = 'abc';
+        $minLength = 3;
+        $blocklist = ['cab', 'abc', 'bca'];
+
+        $sqids = new Sqids($alphabet, $minLength, $blocklist);
+
+        $this->assertSame($minLength, strlen($alphabet));
+        $this->assertSame(count($blocklist), strlen($alphabet));
+
+        $this->expectException(InvalidArgumentException::class);
+        $sqids->encode([0]);
     }
 }