+++ tests/SqidsEncodingTest.php
@@ -22,7 +22,7 @@ class SqidsEncodingTest extends TestCase
         $sqids = new Sqids();
 
         $numbers = [1, 2, 3];
-        $id = '8QRLaD';
+        $id = '86Rf07';
 
         $this->assertSame($id, $sqids->encode($numbers));
         $this->assertSame($numbers, $sqids->decode($id));
@@ -32,7 +32,7 @@ class SqidsEncodingTest extends TestCase
     {
         $sqids = new Sqids();
 
-        $numbers = [0, 0, 0, 1, 2, 3, 100, 1000, 100000, 1000000, $sqids->maxValue()];
+        $numbers = [0, 0, 0, 1, 2, 3, 100, 1000, 100000, 1000000, PHP_INT_MAX];
         $this->assertSame($numbers, $sqids->decode($sqids->encode($numbers)));
     }
 
@@ -41,16 +41,16 @@ class SqidsEncodingTest extends TestCase
         $sqids = new Sqids();
 
         $ids = [
-            'bV' => [0],
-            'U9' => [1],
-            'g8' => [2],
-            'Ez' => [3],
-            'V8' => [4],
-            'ul' => [5],
-            'O3' => [6],
-            'AF' => [7],
-            'ph' => [8],
-            'n8' => [9]
+            'bM' => [0],
+            'Uk' => [1],
+            'gb' => [2],
+            'Ef' => [3],
+            'Vq' => [4],
+            'uw' => [5],
+            'OI' => [6],
+            'AX' => [7],
+            'p6' => [8],
+            'nJ' => [9]
         ];
 
         foreach ($ids as $id => $numbers) {
@@ -64,16 +64,16 @@ class SqidsEncodingTest extends TestCase
         $sqids = new Sqids();
 
         $ids = [
-            'SrIu' => [0, 0],
-            'nZqE' => [0, 1],
-            'tJyf' => [0, 2],
-            'e86S' => [0, 3],
-            'rtC7' => [0, 4],
-            'sQ8R' => [0, 5],
-            'uz2n' => [0, 6],
-            '7Td9' => [0, 7],
-            '3nWE' => [0, 8],
-            'mIxM' => [0, 9]
+            'SvIz' => [0, 0],
+            'n3qa' => [0, 1],
+            'tryF' => [0, 2],
+            'eg6q' => [0, 3],
+            'rSCF' => [0, 4],
+            'sR8x' => [0, 5],
+            'uY2M' => [0, 6],
+            '74dI' => [0, 7],
+            '30WX' => [0, 8],
+            'moxr' => [0, 9]
         ];
 
         foreach ($ids as $id => $numbers) {
@@ -87,16 +87,16 @@ class SqidsEncodingTest extends TestCase
         $sqids = new Sqids();
 
         $ids = [
-            'SrIu' => [0, 0],
-            'nbqh' => [1, 0],
-            't4yj' => [2, 0],
-            'eQ6L' => [3, 0],
-            'r4Cc' => [4, 0],
-            'sL82' => [5, 0],
-            'uo2f' => [6, 0],
-            '7Zdq' => [7, 0],
-            '36Wf' => [8, 0],
-            'm4xT' => [9, 0]
+            'SvIz' => [0, 0],
+            'nWqP' => [1, 0],
+            'tSyw' => [2, 0],
+            'eX68' => [3, 0],
+            'rxCY' => [4, 0],
+            'sV8a' => [5, 0],
+            'uf2K' => [6, 0],
+            '7Cdk' => [7, 0],
+            '3aWP' => [8, 0],
+            'm2xn' => [9, 0]
         ];
 
         foreach ($ids as $id => $numbers) {
@@ -132,18 +132,11 @@ class SqidsEncodingTest extends TestCase
         $this->assertSame([], $sqids->decode('*'));
     }
 
-    public function testDecodingAnInvalidIdWithARepeatingReservedCharacter()
-    {
-        $sqids = new Sqids();
-        $this->assertSame([], $sqids->decode('fff'));
-    }
-
-    public function testEncodeOutOfRangeNumbers()
+    public function testEncodeOutOfRangeNumbersLower()
     {
         $this->expectException(InvalidArgumentException::class);
 
         $sqids = new Sqids();
-        $sqids->encode([$sqids->minValue() - 1]);
-        $sqids->encode([$sqids->maxValue() + 1]);
+        $sqids->encode([-1]);
     }
 }