+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # CHANGELOG
 
+**v0.4.0:** **⚠️ BREAKING CHANGE**
+- **Breaking change**: IDs change. Algorithm has been fine-tuned for better performance [[Issue #11](https://github.com/sqids/sqids-spec/issues/11)]
+- `alphabet` cannot contain multibyte characters
+- `minLength` upper limit has increased from alphabet length to `255`
+- Max blocklist re-encoding attempts has been capped at the length of the alphabet - 1
+- Minimum alphabet length has changed from 5 to 3
+- `minValue()` and `maxValue()` functions have been removed
+- Max integer encoding value is `PHP_INT_MAX`
+
 **v0.3.1:**
 - Bug fix: spec update (PR #7): blocklist filtering in uppercase-only alphabet [[PR #7](https://github.com/sqids/sqids-spec/pull/7)]